[REF] web: move mountWithSearch test helper
This commit moves the test helper `mountWithSearch` to the file
`_framework/search_test_helpers` and re-exports it from the file
`web_test_helpers`

X-original-commit: d5679d26c0a651f387965a3d0d876f21e548824a
