+++ addons/web/static/tests/search/search_bar_menu/comparison_menu.test.js
@@ -2,12 +2,12 @@ import { expect, test } from "@odoo/hoot";
 import { mockDate } from "@odoo/hoot-mock";
 import {
     getFacetTexts,
+    mountWithSearch,
     removeFacet,
     toggleMenuItem,
     toggleMenuItemOption,
     toggleSearchBarMenu,
 } from "@web/../tests/web_test_helpers";
-import { mountWithSearch } from "../helpers";
 import { defineSearchBarModels } from "./models";
 
 import { SearchBarMenu } from "@web/search/search_bar_menu/search_bar_menu";
