+++ addons/web/static/tests/web_test_helpers.js
@@ -95,6 +95,7 @@ export {
     getVisibleButtons,
     isItemSelected,
     isOptionSelected,
+    mountWithSearch,
     openAddCustomFilterDialog,
     pagerNext,
     pagerPrevious,
